Better monospace testing. Further refine setTabstops and analyzeTabstops.
in Elastic tabstops profile dialog, fix sometimes showing wrong default profile and make sure when opening the dialog for a document with a named profile, in case that profile has been changed by opening the dialog in a different document, the change indicator is shown if this profile does not match the saved profile.

Updates to changelog, help and readme.