Default monospace optimization off when using DirectWrite: it doesn't seem to help. Update help accordingly.
Fix a misunderstanding about NPPN_FILECLOSED that caused the DocumentData structure for documents that were still in use to be deleted.

Add a vector for each view that keeps track of which lines already have their custom tabs set. For large rectangular selections this can make redundant calls to setTabstops so fast as to be unnoticeable, which is less error-prone than trying to determine which calls are redundant and suppress them.